Working Control Interface
Got an interface to camera controls working. Still needs to be cleaned
up a bit.

Mostly just ported the controls implementation of pyuvc from cython to
ctypes, with a few minor tweaks to get things working. Their method for
getting and setting values failed on controls that could have negative
values (like Brightness and Hue).

Calls to uvc_get_ctrl_len always fails with an Errno 32 Pipe error. Need
to investigate this.

Some controls I am unable to set. This could be because the camera
doesn't expose them, but I need to investigate by comparing to output of
uvcdynctrl.